Add additional crawler user agents.
Nuzzel is a news aggregation service that goes off of tweets.  If a site is tweeted then nuzzel will attempt to scrape that link- without the crawler user agent it will get non-prerendered html. See http://nuzzel.com/

Discord is a chat bot for gamers. Like Slack, it attempts to expand links that are pasted.  Without detecting the user agent properly it will just show the non-prerendered markup when it expands links. https://discordapp.com/